clone aja semua pak
