+++ src/tools/Connections.java
@@ -31,7 +31,7 @@ public class Connections {
             ods.setServiceName("XE");
 //            ods.setDatabaseName("HR");
             ods.setUser("system");
-            ods.setPassword("ekoeko123");
+            ods.setPassword("pass");
             this.connection = ods.getConnection();
             this.connection.createStatement().execute("alter session set current_schema=hr");
         } catch (Exception e) {
